# docs/swagger.yaml
basePath: /
definitions:
  entitie.Usuario:
    properties:
      email:
        example: teste@email.com
        type: string
      id:
        example: 1
        type: integer
      nome:
        example: teste
        type: string
    type: object
host: localhost
info:
  contact:
    email: italo.ods@hotmail.com
    name: Ítalo Oliveira
    url: https://github.com/dev-ita
  description: simples aplicação CRUD desenvolvida em GO
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Documentação para o Projeto Integrado II
  version: "1.0"
paths:
  /usuarios:
    get:
      consumes:
      - application/json
      description: Retorna todos os detalhes dos usuários
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entitie.Usuario'
            type: array
      summary: Retorna os usuários
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Cria um usuário no banco de dados
      parameters:
      - description: Cria um usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/entitie.Usuario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entitie.Usuario'
      summary: Cria um usuário
      tags:
      - users
  /usuarios/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um usuário correspondente ao ID
      parameters:
      - description: ID of the user to be deleted
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um usuário
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retorna todos os detalhes do usuário através de um ID
      parameters:
      - description: ID of the user
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entitie.Usuario'
      summary: Retorna um único usuário por ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Atualiza um usuário através do Id correspondente
      parameters:
      - description: ID of the user to be updated
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entitie.Usuario'
      summary: Atualiza um usuário
      tags:
      - users
swagger: "2.0"